transfer linked but not working

+++ db/migrate/20141115055318_add_balance_info_to_users.rb
@@ -1,6 +1,6 @@
 class AddBalanceInfoToUsers < ActiveRecord::Migration
   def change
-    add_column :users, :debit, :float
-    add_column :users, :credit, :float
+    add_column :users, :debit, :decimal
+    add_column :users, :credit, :decimal
   end
 end
